requirements.txt
42cc-pystyle==0.0.15
Django==1.6.7
Pillow==3.1.0
South==1.0.2
django-registration-redux==1.2
flake8==2.5.1
mccabe==0.3.1
pep8==1.5.7
pyflakes==1.0.0
selenium==2.48.0
wsgiref==0.1.2
